// frontend/src/app/models/FormXpathScoreMapping.ts
export class FormXpathScoreMapping
{
    formXpathScoreId: number;
    xPath:            null;
    formId:           number;
    maxScore:         number ;
    label:            string;
    type:             null;
    parentXpathId:    number;
    form_meta_id:     number;
    markerClass:      null;
}